backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
python-multipart==0.0.20
python-jose==3.5.0
pydantic-settings==2.11.0
email-validator==2.2.0
